+++ baselines/common/vec_env/subproc_vec_env.py
@@ -1,28 +1,7 @@
-import contextlib
 import multiprocessing as mp
-import os
 
 import numpy as np
-from . import VecEnv, CloudpickleWrapper
-
-@contextlib.contextmanager
-def clear_mpi_env_vars():
-    """
-    from mpi4py import MPI will call MPI_Init by default.  If the child process has MPI environment variables, MPI will think that the child process is an MPI process just like the parent and do bad things such as hang.
-
-    This context manager is a hacky way to clear those environment variables temporarily such as when we are starting multiprocessing
-    Processes.
-    """
-    removed_environment = {}
-    for k, v in list(os.environ.items()):
-        for prefix in ['OMPI_', 'PMI_']:
-            if k.startswith(prefix):
-                removed_environment[k] = v
-                del os.environ[k]
-    try:
-        yield
-    finally:
-        os.environ.update(removed_environment)
+from . import VecEnv, CloudpickleWrapper, clear_mpi_env_vars
 
 ctx = mp.get_context('spawn')
